processing/requirements.txt
celery==4.2.1
numpy==1.15.2
msgpack-python==0.5.6
msgpack-numpy==0.4.4.1
python-dotenv==0.9.1

dash==0.28.1
dash-core-components==0.33.0
dash-html-components==0.13.2

flask==1.0.2
flask-caching==1.4.0

redis==2.10.6

lz4==2.1.0
